+++ src/commands.js
@@ -7,13 +7,13 @@ const chokidar = require('chokidar');
 const { join: pathJoin, resolve: pathResolve, basename: pathBasename } = require('path');
 
 const utils = require('./utils');
+const content = require('./content');
 const pkg = require('../package.json');
-const { libFile, createGdnlibFile } = require('./content');
 
 const log = console.log;
 
 /**
- * Define the targets that can be specified during the `new` command and the default targets if none are specified.
+ * The targets that can be used with the `new` command's `--targets` option.
  */
 const validTargets = ['windows', 'linux', 'osx'];
 
@@ -22,41 +22,52 @@ const validTargets = ['windows', 'linux', 'osx'];
  */
 module.exports = {
   /**
-   * The `new` command is used to create an environment for your Rust modules. The environment is created as a directory in the specified destination.
+   * The `new` command is used to create the library that will contain your Rust modules.
    * 
-   * @param {string} destination The destination directory for the environment.
-   * @param {string} godotProjectDir The directory of the Godot project that the Rust modules are for.
-   * @param {string} [targets] The build targets that should be set.
+   * The name of the library that will contain your Rust modules. The name of the library is recommended to be the same or similar in name to your game. 
+   * Also keep in mind that the library is created using `cargo new` so you should abide by the cargo project naming standards.
+   * 
+   * @async
+   * 
+   * @param {string} name The name of the library that will contain your Rust modules. The name of the library is recommended to be the same name as your game, snake_case, maybe with `_modules` at the end. Also keep in mind that the library is created using `cargo new`
+   * @param {string} godotProjectDir The directory that contains the project.godot file of the game that the modules are for.
+   * @param {string} [targets] The build targets that should be set. As of writing this, the available targets are windows, linux, and osx with the default being just windows.
    */
-  new(destination, godotProjectDir, targets = 'windows') {
-    /**
-     * First, we check to see if the destination directory for the environment exists.
-     * 
-     * If it does, we let the user know and then just stop the script.
-     */
-    if (fs.pathExistsSync(destination)) {
-      console.log('The destination folder already exists, please choose another destination for the environment.');
+  async new(name, godotProjectDir, targets = 'windows') {
+    log(chalk.white('creating library'));
+    // Check to see if the library already exists in the current directory.
+    if (fs.pathExistsSync(name)) {
+      log(chalk.red('A library with the specified name already exists, please choose another name for the library.'));
       return;
     }
 
-    /**
-     * Next, we check to see if the provided Godot project is valid. In order for a Godot project to be considered valid, it needs
-     * to have a project.godot file.
-     * 
-     * If there is no godot.project file, then we let the user know and stop the script.
-     */
+    // Check to see if the path to the Godot project is valid. In order it to be considered valid, it needs to have a project.godot file.
     if (!fs.pathExistsSync(pathJoin(godotProjectDir, 'project.godot'))) {
-      console.log('The godot project dir provided is not valid.');
+      log(chalk.red('The godot project dir provided is not valid.'));
       return;
     }
 
-    /**
-     * So the destination directory can be created and the Godot project directory is valid so we create the destination directory.
-     * 
-     * Also create the 'rust-modules' directory in the Godot project so that we don't clutter the root directory.
-     */
-    fs.mkdirpSync(destination);
-    fs.mkdirpSync(pathJoin(godotProjectDir, 'rust-modules'));
+    // Run the `cargo new` command in the current directory to create the library.
+    try {
+      await shelljs.exec(`cargo new ${name} --lib`);
+    } catch (err) {
+      log(chalk.red(err));
+      log(chalk.red('Could not create the library, check the details of the error above and try running the command again.'));
+      return;
+    }
+
+    // We don't want to overwrite the contents of the Cargo.toml file because of entries like author so we're going to modify just the pieces of it that we need to.
+    const moduleToml = fs.readFileSync(`${name}/Cargo.toml`, { encoding: 'utf-8' }).split('\n');
+    if (!moduleToml.includes('[lib]')) {
+      const libInsertPoint = moduleToml.indexOf('[dependencies]');
+
+      moduleToml.splice(libInsertPoint, 0, '[lib]');
+      moduleToml.splice(libInsertPoint + 1, 0, '');
+      moduleToml.splice(libInsertPoint + 1, 0, 'crate-type = ["cdylib"]');
+    }
+    const depInsertPoint = moduleToml.indexOf('[dependencies]');
+    moduleToml.splice(depInsertPoint + 1, 0, 'gdnative = { git = "https://github.com/GodotNativeTools/godot-rust" }');
+    fs.writeFileSync(`${name}/Cargo.toml`, moduleToml.join('\n'));
 
     /**
      * Check to see if the provided targets are valid.
@@ -73,230 +84,174 @@ module.exports = {
       }
     });
 
-    /**
-     * Create the config file that contains the path to the Godot project and the targets that should be set.
-     */
-    const config = { godotProjectDir: pathResolve(godotProjectDir), targets, modules: [] };
-    fs.outputJsonSync(pathJoin(destination, 'godot-rust-helper.json'), config);
+    // Define the contents of the config file and write it to the library directory.
+    const config = {
+      name: pathBasename(name),
+      godotProjectDir: pathResolve(godotProjectDir),
+      targets,
+      modules: []
+    };
+
+    // Save the config file as we're done working with it for now.
+    fs.outputJsonSync(`${name}/godot-rust-helper.json`, config);
+
+    // Create the initial src/lib.rs file.
+    fs.writeFileSync(`${name}/src/lib.rs`, content.createInitialLibFile());
+
+    // Create the 'rust-modules' directory in the Godot project so that we don't clutter the root directory.
+    fs.mkdirpSync(pathJoin(godotProjectDir, 'rust-modules'));
+
+    // Create the gdnlib file for the module and save it to the Godot project directory.
+    const gdnlib = content.createGdnlibFile(name, config.targets);
+    fs.writeFileSync(`${config.godotProjectDir}/rust-modules/${pathBasename(name)}.gdnlib`, gdnlib);
+    log(chalk.green('library created successfully'));
   },
 
   /**
-   * The `create` command is used inside of an environment created with `new` and is used to initialize a new Rust module.
+   * The `create` command is used inside of a library created with `new` and is used to initialize a new Rust module for use in the game.
+   * 
+   * The name passed to this command should be the class name of the module. Class names must start with capital letters. Examples include 'Player', 'Princess', 'Mob', 'HUD', etc.
    * 
    * @async
    * 
-   * @param {string} name The name of the Rust module to create.
+   * @param {string} name The class name of the module to create; examples include 'Player', 'Princess', 'Mob', 'HUD', etc.
    */
   async create(name) {
-    /**
-     * First, we have to make sure that we are in an environment created by `new`. This is done by checking to see if there is a 
-     * godot-rust-helper.json configuration file present.
-     * 
-     * If we are not in an environment, we let the user know and stop the script.
-     */
+    log(chalk.white('creating module'));
+
+    // Check to see if we are in a library created with the `new` command which is done by checking to see if there is a godot-rust-helper.json config file present.
     if (!fs.pathExistsSync('godot-rust-helper.json')) {
-      console.log('This command can only be used inside of an envrionment created with the new command');
+      log(chalk.red('This command can only be used inside of a library created with the new command'));
       return;
     }
 
-    /**
-     * Next, we have to check if a module with the same name already exists.
-     * 
-     * If it does, then we let the user know and stop the script.
-     */
-    if (fs.pathExistsSync(name)) {
-      console.log('A module with the same name already exists.');
-      return;
-    }
-
-    /**
-     * We are in an envrionment and the module does not already exist, we can now create the module.
-     * 
-     * We do this by running `cargo new --lib` to create a new cargo project.
-     */
-    try {
-      await shelljs.exec(`cargo new ${name} --lib`);
-    } catch (err) {
-      console.log(err);
-      console.log('Could not create Rust module, try running the command again.');
-      return;
-    }
-
-    /**
-    * Check if the Cargo.toml file has everything we need and if not then we have to add it.
-    */
-    const moduleToml = fs.readFileSync(`${name}/Cargo.toml`, { encoding: 'utf-8' }).split('\n');
-
-    if (!moduleToml.includes('[lib]')) {
-      const libInsertPoint = moduleToml.indexOf('[dependencies]');
-
-      moduleToml.splice(libInsertPoint, 0, '[lib]');
-      moduleToml.splice(libInsertPoint + 1, 0, '');
-      moduleToml.splice(libInsertPoint + 1, 0, 'crate-type = ["cdylib"]');
-    }
-
-    const depInsertPoint = moduleToml.indexOf('[dependencies]');
-    moduleToml.splice(depInsertPoint + 1, 0, 'gdnative = { git = "https://github.com/GodotNativeTools/godot-rust" }');
-
-    fs.writeFileSync(`${name}/Cargo.toml`, moduleToml.join('\n'));
-
-    /**
-     * Replace the contents of the default src/lib.rs with the example HelloWorld module.
-     */
-    fs.writeFileSync(`${name}/src/lib.rs`, libFile);
-
-    /**
-     * Save the module name to the config file so that it can be destroyed later if needed.
-     */
+    // Check to see if a module with the same name was already created.
+    const nameNormalized = name.toLowerCase();
     const config = fs.readJsonSync('godot-rust-helper.json');
+    if (config.modules.includes(name)) {
+      log(chalk.red('A module with the same name already exists.'));
+      return;
+    }
+
+    // Save the module name to the config file so that it can be worked with later.
     config.modules.push(name);
-
-    /**
-     * Create the folder for the module in the Godot project directory.
-     */
-    fs.ensureDirSync(`${config.godotProjectDir}/rust-modules/${name}`);
-
-    const gdnlib = createGdnlibFile(name, config.targets);
-    fs.writeFileSync(`${config.godotProjectDir}/rust-modules/${name}/${name}.gdnlib`, gdnlib);
-
-    /**
-     * Lastly write to the config file to save the module created.
-     */
     fs.outputJsonSync('godot-rust-helper.json', config);
+
+    // Create a new src/lib.rs with the new module added to it.
+    const libFileNew = content.createLibFile(config.modules);
+    fs.writeFileSync(`src/lib.rs`, libFileNew);
+
+    // Create a sample lib file for the module that prints "Hello, World!" when the game is run.
+    fs.writeFileSync(`src/${nameNormalized}.rs`, content.createModuleFile(name));
+
+    log(chalk.green('module created'));
   },
 
   /**
-   * The `destroy` command is used inside of an environment created with `new` and is used to remove a Rust module created with `create`.
+   * The `destroy` command is used inside of a library created with `new` and is used to remove a module created with `create`.
    * 
-   * @param {string} name The name of the Rust module to destroy.
+   * @param {string} name The name of the module to destroy. This should be the same name provided when it was created.
    */
   destroy(name) {
+    log(chalk.white('destroying module'));
+
+    // Check to see if we are in a library directory as the command can only be run from inside there.
+    if (!fs.pathExistsSync('godot-rust-helper.json')) {
+      log(chalk.red('This command can only be used inside of a library created with the new command.'));
+      return;
+    }
+
+    // Check to see if the file actually has a .rs file in the src directory.
+    const nameNormalized = name.toLowerCase();
+    if (!fs.pathExistsSync(`src/${nameNormalized}.rs`)) {
+      log(chalk.red('The module to delete does not exist in the src directory.'));
+      return;
+    }
+
+    // Get the config file so we can see what modules have been created.
     const config = fs.readJsonSync('godot-rust-helper.json');
 
-    /**
-     * First, we have to make sure that we are in an environment created by `new`. This is done by checking to see if there is a 
-     * godot-rust-helper.json configuration file present.
-     * 
-     * If we are not in an environment, we let the user know and stop the script.
-     */
-    if (!fs.pathExistsSync('godot-rust-helper.json')) {
-      console.log('This command can only be used inside of an envrionment created with the new command');
-      return;
-    }
-
-    /**
-     * Next, we have to check if the module to delete actually exists.
-     * 
-     * If it does not, then we let the user know and stop the script.
-     */
-    if (!fs.pathExistsSync(name)) {
-      console.log('The module to delete does not exist');
-      return;
-    }
-
-    /**
-     * Everything's good so we remove the module from the config, from the environment, and from the Godot project.
-     */
-    config.modules = config.modules.filter(module => module !== name);
+    // Remove the module from the config file and save it again.
+    config.modules = config.modules.filter(module => module.toLowerCase() !== nameNormalized);
     fs.outputJsonSync('godot-rust-helper.json', config);
 
-    fs.removeSync(name);
+    // Create a new src/lib.rs file based on the modules that are left over.
+    let libFile;
+    if (config.modules.length === 0) libFile = content.createInitialLibFile();
+    else libFile = content.createLibFile(config.modules);
+    fs.writeFileSync(`src/lib.rs`, libFile);
 
-    fs.removeSync(`${config.godotProjectDir}/rust-modules/${name}`);
+    // Remove the module's .rs file.
+    fs.removeSync(`src/${nameNormalized}.rs`);
+
+    log(chalk.green('module destroyed'));
   },
 
   /**
-   * The `import` command is used to import a Rust module from outside the environment.
+   * The `import` command is used to import a Rust module from a another library into this one.
    * 
-   * This command performs a basic validation of the module and then moves the module to the environment's directory
-   * 
-   * @param {string} pathToModule The path to the Rust module to import.
+   * @param {string} path The path to the library to import the module from.
+   * @param {string} name The name of the module to import. This should be the same name provided when it was created.
    */
-  import(pathToModule) {
+  import(path, name) {
     log(chalk.white('starting import'));
-    /**
-     * We want this command to only be usable from the root of the environment's directory.
-     * 
-     * If we look for the closest config file we could run into issues with other environments so we'll enforce this as it makes sense anyways.
-     */
+
+    // Check to see if we are in a library.
     if (!fs.pathExistsSync('godot-rust-helper.json')) {
-      console.log('The import command can only be used from the root of the environment.');
+      log(chalk.red('The import command can only be used from the root of the library.'));
       return;
     }
 
-    /**
-     * We're in the environment root, now we check to see if the path is valid.
-     */
-    if (!fs.pathExistsSync(pathToModule)) {
-      console.log('The path to the Rust module to import does not exist.');
+    // Check to see if the path to the library to import the module from is valid.
+    if (!fs.pathExistsSync(path)) {
+      log(chalk.red('There path to the library that contains the module to import is not valid.'));
       return;
     }
 
-    /**
-     * Next, we perform a basic check to make sure that it is a Rust module (contains a cargo.toml file).
-     */
-    if (!fs.pathExistsSync(pathJoin(pathToModule, 'Cargo.toml'))) {
-      console.log('The Rust module provided does not contain a Cargo.toml file and cannot be imported.');
+    // Check to see if the library to import the module from is a valid library.
+    if (!fs.pathExistsSync(pathJoin(path, 'godot-rust-helper.json'))) {
+      log(chalk.red('The path to the library provided is not valid.'));
       return;
     }
 
-    /**
-     * Now that we know it exists and is semi-valid, we check the full validity of the module.
-     * 
-     * This checks to make sure that the Cargo.toml contains the basic necessary components of a Rust module.
-     */
-    const moduleToml = fs.readFileSync(pathJoin(pathToModule, 'Cargo.toml'), { encoding: 'utf-8' });
-    if (!moduleToml.includes('cdylib') || !moduleToml.includes('gdnative')) {
-      console.log('The Rust module is missing key aspects of a basic Rust module such as the gdnative dependency and cannot be imported.');
+    // Check to see if the module to import exists in the target's config.
+    const importConfig = fs.readJSONSync(pathJoin(path, 'godot-rust-helper.json'));
+    if (!importConfig.modules.includes(name)) {
+      log(chalk.red('The module to import does not exist in the library\'s config.'));
       return;
     }
 
-    /**
-     * Finally we copy the module into the environment root, create the gdnlib file for it if it does not exist, and add it to the 
-     * godot-rust-helper.json config so that it can be kept track of.
-     * 
-     * Note: The gdnlib process also exists in the `create` command so it should be made into a reusable function.
-     */
-    log(chalk.white('creating gdnlib file...'));
-
-    const moduleName = pathBasename(pathToModule);
-    fs.copySync(pathToModule, pathJoin(process.cwd(), moduleName));
-
+    // Add the module to import to the config.
     const config = fs.readJsonSync('godot-rust-helper.json');
-    config.modules.push(pathBasename(pathToModule));
+    config.modules.push(name);
+    fs.outputJSONSync('godot-rust-helper.json', config);
 
-    fs.ensureDirSync(`${config.godotProjectDir}/rust-modules/${moduleName}`);
+    // Copy the module into the current library.
+    log(chalk.white('importing module...'));
+    const namelc = name.toLowerCase();
+    fs.copyFileSync(pathJoin(path, 'src', `${namelc}.rs`), `src/${namelc}.rs`);
 
-    const gdnlib = createGdnlibFile(moduleName, config.targets);
-    fs.writeFileSync(`${config.godotProjectDir}/rust-modules/${moduleName}/${moduleName}.gdnlib`, gdnlib);
-
-    fs.outputJsonSync('godot-rust-helper.json', config);
-
+    // Add its entry to the src/lib.rs file.
+    fs.writeFileSync('src/lib.rs', content.createLibFile(config.modules));
     log(chalk.green('import complete'));
   },
 
   /**
-   * Runs the build command and copies the target files into the Godot project directory.
+   * Runs `cargo build` and copies the target files into the Godot project directory.
    * 
    * @async
    */
   async build() {
     log(chalk.white.underline(`godot-rust-helper v${pkg.version}`));
     log(chalk.cyan('building...'));
-    
-    /**
-     * First we have to find the root of the module, because it might not be where the command is being run from.
-     */
+
+    // Find the root of the module, aka the directory that contains the Cargo.toml.
     const moduleDir = utils.findFile(process.cwd(), 'Cargo.toml');
 
-    /**
-     * Get the config so that we can check the targets later.
-     */
-    const config = fs.readJsonSync(pathResolve(moduleDir, '..', 'godot-rust-helper.json'));
+    // Get the config so that we can check the targets later.
+    const config = fs.readJsonSync(pathJoin(utils.findFile(process.cwd(), 'godot-rust-helper.json'), 'godot-rust-helper.json'));
 
-    /**
-     * Then we have to run Cargo build to create the build files.
-     */
+    // Run the `cargo build` command to generate the target files.
     try {
       shelljs.exec(`cargo build`);
     } catch (err) {
@@ -304,20 +259,10 @@ module.exports = {
       return;
     }
 
-    /**
-     * Get the name of the module from the base path to it.
-     */
-    const isWindows = process.platform == 'win32'
-    const name = moduleDir.split( isWindows ? '\\' : '/').pop();
-
-    /**
-     * Get the base path to the build files.
-     */
+    // Get the path to where the build files are stored.
     const base = pathJoin(moduleDir, 'target', 'debug');
 
-    /**
-     * Go through the targets specified in the config file and copy the build files to the Godot project dir.
-     */
+    // Go through each of the targets specified in the config and copy the matching build files to the Godot project dir.
     config.targets.map(target => {
       const normalized = {
         linux: 'so',
@@ -325,9 +270,10 @@ module.exports = {
         windows: 'dll'
       };
 
-      const file = pathJoin(base, `${isWindows ? '' : `lib`}${name}.${normalized[target]}`);
+      const isWindows = process.platform == 'win32'
+      const file = pathJoin(base, `${isWindows ? '' : `lib`}${config.name}.${normalized[target]}`);
 
-      shelljs.cp(file, pathJoin(config.godotProjectDir, 'rust-modules', name));
+      shelljs.cp(file, pathJoin(config.godotProjectDir, 'rust-modules'));
 
       log(chalk.green('build complete'));
     });